refactor: simplified yet production standard full stack typesafety
