+++ apps/server/src/schemas/index.ts
@@ -1,0 +1,5 @@
+export * from "./admin";
+export * from "./anime";
+export * from "./common";
+export * from "./library";
+export * from "./users";
