+++ apps/web/package.json
@@ -18,6 +18,7 @@
     "better-auth": "catalog:",
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
+    "elysia": "^1.4.21",
     "framer-motion": "^12.34.0",
     "lucide-react": "^0.564.0",
     "next": "^16.0.10",
@@ -26,17 +27,16 @@
     "react": "19.2.3",
     "react-dom": "19.2.3",
     "react-youtube": "^10.1.0",
+    "server": "workspace:*",
     "sonner": "^2.0.5",
     "tailwind-merge": "^3.3.1",
     "tw-animate-css": "^1.3.4",
-    "zod": "catalog:",
-    "server": "workspace:*",
-    "elysia": "^1.4.21"
+    "zod": "catalog:"
   },
   "devDependencies": {
-    "@types/bun": "catalog:",
     "@tailwindcss/postcss": "^4.1.10",
     "@tanstack/react-query-devtools": "^5.91.1",
+    "@types/bun": "catalog:",
     "@types/node": "^20",
     "@types/react": "19.2.7",
     "@types/react-dom": "19.2.3",
