--- packages/api/src/contracts/common.ts
@@ -1,3 +1,0 @@
-export type ApiErrorResponse = {
-  error: string;
-};
