+++ packages/api/src/index.ts
@@ -1,7 +1,3 @@
 export * from "./services/anime.service";
 export * from "./services/library.service";
 export * from "./services/user.service";
-export * from "./contracts/anime";
-export * from "./contracts/library";
-export * from "./contracts/users";
-export * from "./contracts/common";
